UIE-93 Decouple data clients - useWorkspaces - PR feedback, including breaking out useLoadedDataEvents to it's own module file. - improved jsdoc for useCachedData.
